+++ bq_scripts/pull_gcp_bq_as_client.py
