+++ bq_scripts/rd_gcs_bq.py
