+++ create_report/requirements.txt
@@ -1,0 +1,5 @@
+google-cloud-storage
+matplotlib
+numpy
+pandas
+seaborn
