+++ create_report/src/main.py
@@ -1,14 +1,78 @@
+# -*- coding: utf-8 -*-
 """
-Create a report file for emailing to participants
+Created on Wed May 29 13:52:24 2019
+script to build a week long water quality data
+report and send it as a Gmail attachment
+Read the token built by quickstart.py from credentials.json
+Follow tutorial here https://developers.google.com/gmail/api/quickstart/python
+to enable Google Gmail API and get credentials.json file
+
+This script can be run as a cron job:
+5 7 * * * conda activate gcp-thingspeak;python /home/jf/Documents/projects/thermoflux-particle-gcp-python/email_report_gcf.py >> /home/jf/Documents/projects/thermoflux-particle-gcp-python/cron.log 2>&1;conda deactivate
+
+@author: John Franco Saraceno
 """
+# import necessary packages
+import base64
+from datetime import date
+from io import BytesIO, StringIO
+import os
+import sys
+from google.cloud import storage
+import matplotlib.pyplot as plt
+import matplotlib.dates as mdates
+import numpy as np
+import pandas as pd
+import seaborn as sns
+
+# set plotting format
+days = mdates.DayLocator(interval=1)  # every year
+hours = mdates.HourLocator(interval=6)  # every month
+date_form = mdates.DateFormatter("%m/%d")
+# colors
+FLATUI = ["#9b59b6", "#3498db", "#95a5a6", "#e74c3c", "#34495e", "#2ecc71", "#f4cae4"]
+
+sns.set(font_scale=2, style="whitegrid")
+
+plt.rcParams.update({"axes.facecolor": "snow"})
+plt.rcParams["xtick.major.size"] = 12
+plt.rcParams["ytick.major.size"] = 12
+plt.rcParams["ytick.major.width"] = 1
+plt.rcParams["xtick.major.width"] = 1
+plt.rcParams["xtick.bottom"] = True
+plt.rcParams["ytick.left"] = True
 
 
-def make_a_plot(dataframe, figname, savefig=True):
+def upload_blob(bucket_name, source_file_name, destination_blob_name, storage_client):
+    """Uploads a file to the bucket."""
+    # bucket_name = "your-bucket-name"
+    # source_file_name = "local/path/to/file"
+    # destination_blob_name = "storage-object-name"
+    bucket = storage_client.bucket(bucket_name)
+    blob = bucket.blob(destination_blob_name)
+    blob.upload_from_filename(source_file_name)
+    print("File {} uploaded to {}.".format(source_file_name, destination_blob_name))
+    return
+
+
+def fetch_bucket(bucket_name, bucket_filename):
+    """Function to fetch a gcs storage
+    object and return a pandas dataframe"""
+    bucket = storage_client.get_bucket(bucket_name)
+    blob = storage.blob.Blob(bucket_filename, bucket)
+    content = blob.download_as_string()
+    dfbytes = pd.read_csv(
+        BytesIO(content), na_values=["NAN", "nan", -999], index_col="Datetime (PST)"
+    )
+    dfbytes.index = pd.to_datetime(dfbytes.index)
+    return dfbytes
+
+
+def make_a_plot(df, figname, savefig=True):
     """Function to plot data as stacked time series plots"""
     # make the datetimeindex timezone naive to correctly work with plot ticks
-    df = dataframe.copy()
     df.index = df.index.tz_localize(None)
-    # df = df["02-26-2021":].copy()
+    df = df.copy()
     plot_cols = df.columns
     n_plot_cols = 4
     # colors = plt.cm.tab10(np.linspace(0, 1, n_plot_cols))
@@ -40,7 +104,9 @@ def make_a_plot(dataframe, figname, savefig=True):
         fontweight="bold",
     )
 
-    axes[0].plot(df.index.values, df["Battery Voltage"], marker="o", color="tab:blue")
+    axes[0].plot(
+        df.index.values, df["Battery voltage (V)"], marker="o", color="tab:blue"
+    )
     # do some axis operations for each subplot
     # add titles
     # add axis labels
@@ -60,7 +126,7 @@ def make_a_plot(dataframe, figname, savefig=True):
     # plot the data
     axes[1].plot(
         df.index.values,
-        df["Temperature"],
+        df["Temperature (C)"],
         marker="^",
         color="tab:orange",
     )
@@ -77,7 +143,9 @@ def make_a_plot(dataframe, figname, savefig=True):
 
     axes[1].grid(color="k", ls="--", lw=1.25)
     # plot the data
-    axes[2].plot(df.index.values, df["Net Radiation"], marker="o", color="tab:cyan")
+    axes[2].plot(
+        df.index.values, df["Net Radiation (W/m2)"], marker="o", color="tab:cyan"
+    )
     # format the ticks
     axes[2].xaxis.set_major_locator(days)
     axes[2].xaxis.set_major_formatter(date_form)
@@ -90,7 +158,7 @@ def make_a_plot(dataframe, figname, savefig=True):
     axes[2].grid(color="k", ls="--", lw=1.25)
     # plot the data
     axes[3].plot(
-        *splitSerToArr(df["Sensible Heat Flux"].dropna()),
+        *splitSerToArr(df["Sensible heat flux (W/m2)"].dropna()),
         marker="o",
         markersize=6,
         ls="-",
@@ -126,3 +194,34 @@ def make_a_plot(dataframe, figname, savefig=True):
     if savefig:
         plt.savefig(figname, format=image_fmt)
     return
+
+
+def splitSerToArr(ser):
+    return [ser.index, ser.values]
+
+
+def create_report(event, context):
+    bucket_name = os.environ["BUCKET_NAME"]  # "thermoflux-output"
+    bucket_filename_read = os.environ[
+        "BUCKET_FILENAME_READ"
+    ]  # "alfalfa_demo_table_output.csv"
+    reports_bucket_name = os.environ["REPORTS_BUCKET_NAME"]  # "thermoflux-reports"
+    report_filename = os.environ["REPORT_FILENAME"]  # "report.png"
+    # bucket_name = "thermoflux-output"
+    # bucket_filename_read = "alfalfa_demo_table_output.csv"
+    # reports_bucket_name = "thermoflux-reports"
+    # report_filename = "report.png"
+    sites = ["LT_MicroIQ_Alfalfa"]
+    today = date.today().strftime("%m/%d/%Y")
+
+    storage_client = storage.Client()
+    for SITE in sites:
+        print(SITE)
+        # create the output figure name
+        FIGNAME = SITE + " - Daily" + ".png"
+        data = fetch_bucket(bucket_name, bucket_filename_read)
+        # plot the data and save as an image
+        make_a_plot(data, FIGNAME)
+        # upload the report to the report gcs data bucket
+        upload_blob(reports_bucket_name, FIGNAME, report_filename, storage_client)
+    print("Done!")
