+++ format_gcs_datafile/cloudbuild.yaml
@@ -1,0 +1,4 @@
+steps:
+  - name: "gcr.io/google.com/cloudsdktool/cloud-sdk"
+    args: ["./scripts/deploy-gcs.sh"]
+    waitFor: ["-"]
