+++ format_gcs_datafile/requirements.txt
@@ -1,0 +1,2 @@
+google-cloud-storage
+pandas
