+++ format_gcs_datafile/scripts/.env.local
@@ -1,0 +1,2 @@
+export FUNCTION_NAME=format_gcs_datafile
+export BUCKET_NAME=thermoflux-bq-data
