+++ format_gcs_datafile/scripts/deploy-gcs.sh
@@ -1,0 +1,13 @@
+#! /bin/bash
+
+DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+SOURCE_DIR="${DIR}/../src"
+
+source "${DIR}/.env.local"
+
+gcloud functions \
+  deploy ${FUNCTION_NAME} \
+  --source=${SOURCE_DIR} \
+  --runtime=python38 \
+  --trigger-resource{${BUCKET_NAME}} \
+  --trigger-event google.storage.object.finalize
